Rename substitute_nodes_with_node to contract_nodes.
