+++ src/graph.rs
@@ -1484,33 +1484,33 @@ impl PyGraph {
     ///     will always appear as the target of its edges as returned by
     ///     :meth:`~retworkx.PyGraph.edge_list`.
     ///
-    /// :param list to_replace: A set of nodes to be removed and replaced
+    /// :param list nodes: A set of nodes to be removed and replaced
     ///     by the new node. Any nodes not in the graph are ignored.
     ///     If empty, this method behaves like :meth:`PyDiGraph.add_node`
     ///     (but slower).
     /// :param object obj: The data/weight to associate with the new node.
     /// :param weight_combo_fn: An optional python callable that, when
     ///     specified, is used to merge parallel edges introduced by the
-    ///     substitution, which will occur when multiple nodes in
-    ///     ``to_replace`` have an incoming edge
+    ///     contraction, which will occur when multiple nodes in
+    ///     ``nodes`` have an incoming edge
     ///     from the same source node or when multiple nodes in
-    ///     ``to_replace`` have an outgoing edge to the same target node.
+    ///     ``nodes`` have an outgoing edge to the same target node.
     ///     If this instance of :class:`retworkx.PyDiGraph` is a multigraph,
     ///     leave this unspecified to preserve parallel edges. If unspecified
     ///     when not a multigraph, parallel edges and their weights will be
     ///     combined by choosing one of the edge's weights arbitrarily based
     ///     on an internal iteration order, subject to change.
     /// :returns: The index of the newly created node.
-    #[pyo3(text_signature = "(self, to_replace, obj, /, weight_combo_fn=None)")]
-    pub fn substitute_nodes_with_node(
+    #[pyo3(text_signature = "(self, nodes, obj, /, weight_combo_fn=None)")]
+    pub fn contract_nodes(
         &mut self,
         py: Python,
-        to_replace: Vec<usize>,
+        nodes: Vec<usize>,
         obj: PyObject,
         weight_combo_fn: Option<PyObject>,
     ) -> PyResult<usize> {
         let mut indices_to_remove: IndexSet<NodeIndex, ahash::RandomState> =
-            to_replace.into_iter().map(NodeIndex::new).collect();
+            nodes.into_iter().map(NodeIndex::new).collect();
 
         // Create new node.
         let node_index = self.graph.add_node(obj);
